requirements.txt
requests
plotly
kaleido
clickhouse-driver
pandas
numpy
pillow
opencv-python
matplotlib
pyspark[sql]
findspark
